# example/openapi/api-docs.yaml
openapi: 3.0.1
info:
  title: MMS Example API
  description: Documentation for MMS API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.txt
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- basicAuth: []
- bearerToken: []
paths:
...
  /projects/{projectId}/refs/{refId}/views:
    put:
      tags:
      - Views
      operationId: getViews
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementsRequest'
        required: true
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
    post:
      tags:
      - Views
      operationId: createOrUpdateViews
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementsRequest'
        required: true
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
  /projects/{projectId}/refs/{refId}/views/{viewId}:
    get:
      tags:
      - Views
      operationId: getView
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      - name: viewId
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
  /projects/{projectId}/refs/{refId}/groups:
    get:
      tags:
      - Views
      operationId: getGroups
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupsResponse'
  /projects/{projectId}/refs/{refId}/mounts:
    get:
      tags:
      - Views
      operationId: getMounts
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MountsResponse'
  /projects/{projectId}/refs/{refId}/documents:
    get:
      tags:
      - Views
      operationId: getDocuments
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentsResponse'
...
  /projects/{projectId}/refs/{refId}:
    get:
      tags:
      - Refs
      operationId: getRef
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefsResponse'
    delete:
      tags:
      - Refs
      operationId: deleteRef
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefsResponse'
...
  /projects/{projectId}/refs/{refId}/elements:
    get:
      tags:
      - Elements
      operationId: getAllElements
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      - name: commitId
        in: query
        required: false
        schema:
          type: string
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
    put:
      tags:
      - Elements
      operationId: getElements
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      - name: commitId
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementsRequest'
        required: true
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
    post:
      tags:
      - Elements
      operationId: createOrUpdateElements
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementsRequest'
        required: true
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
    delete:
      tags:
      - Elements
      operationId: deleteElements
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: refId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementsRequest'
        required: true
      responses:
        200:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
...
components:
  schemas:
...
    ElementJson:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        type:
          type: string
        _inRefIds:
          type: array
          readOnly: true
          items:
            type: string
            readOnly: true
        _projectId:
          type: string
          readOnly: true
        _refId:
          type: string
          readOnly: true
        _creator:
          type: string
          readOnly: true
        _docId:
          type: string
          readOnly: true
        _modifier:
          type: string
          readOnly: true
        _created:
          type: string
          readOnly: true
        _commitId:
          type: string
          readOnly: true
        _modified:
          type: string
          readOnly: true
      additionalProperties: true
    ElementsRequest:
      required:
      - elements
      type: object
      properties:
        source:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        elements:
          type: array
          items:
            $ref: '#/components/schemas/ElementJson'
    ElementsResponse:
      required:
      - elements
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        elements:
          type: array
          items:
            $ref: '#/components/schemas/ElementJson'
    Rejection:
      type: object
      properties:
        object:
          type: object
        code:
          type: integer
          format: int32
        message:
          type: string
      nullable: true
    GroupsResponse:
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        groups:
          type: array
          items:
            $ref: '#/components/schemas/ElementJson'
    MountJson:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        type:
          type: string
        _projectId:
          type: string
          readOnly: true
        _refId:
          type: string
          readOnly: true
        _creator:
          type: string
          readOnly: true
        _docId:
          type: string
          readOnly: true
        _modifier:
          type: string
          readOnly: true
        _created:
          type: string
          readOnly: true
        _commitId:
          type: string
          readOnly: true
        _modified:
          type: string
          readOnly: true
      additionalProperties: true
    MountsResponse:
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        projects:
          type: array
          items:
            $ref: '#/components/schemas/MountJson'
    DocumentsResponse:
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/ElementJson'
...
    RefJson:
      required:
      - name
      - type
      type: object
      properties:
        status:
          type: string
          readOnly: true
        description:
          type: string
        parentRefId:
          type: string
          default: master
        deleted:
          type: boolean
          readOnly: true
        parentCommitId:
          type: string
        name:
          type: string
        id:
          type: string
        type:
          type: string
          enum:
          - Branch
          - Tag
        _projectId:
          type: string
          readOnly: true
        _refId:
          type: string
          readOnly: true
        _creator:
          type: string
          readOnly: true
        _docId:
          type: string
          readOnly: true
        _modifier:
          type: string
          readOnly: true
        _created:
          type: string
          readOnly: true
        _modified:
          type: string
          readOnly: true
      additionalProperties: true
    RefsResponse:
      required:
      - refs
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        refs:
          type: array
          items:
            $ref: '#/components/schemas/RefJson'
...
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic